+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/parser/PluginParserTest.kt
@@ -1170,6 +1170,42 @@ private class PluginParserTest : ParserTestCase() {
                 val actual = parseResource("tests/parser/marklogic-7.0/ElementDeclTest_MissingClosingParenthesis.xq")
                 assertThat(prettyPrintASTNode(actual), `is`(expected))
             }
+
+            @Nested
+            @DisplayName("XQuery 1.0 EBNF (134) ElementNameOrWildcard")
+            internal inner class ElementNameOrWildcard {
+                @Test
+                @DisplayName("element name")
+                fun elementName() {
+                    val expected = loadResource("tests/parser/marklogic-7.0/ElementDeclTest_ElementName.txt")
+                    val actual = parseResource("tests/parser/marklogic-7.0/ElementDeclTest_ElementName.xq")
+                    assertThat(prettyPrintASTNode(actual), `is`(expected))
+                }
+
+                @Test
+                @DisplayName("element name; compact whitespace")
+                fun elementName_compactWhitespace() {
+                    val expected = loadResource("tests/parser/marklogic-7.0/ElementDeclTest_ElementName_CompactWhitespace.txt")
+                    val actual = parseResource("tests/parser/marklogic-7.0/ElementDeclTest_ElementName_CompactWhitespace.xq")
+                    assertThat(prettyPrintASTNode(actual), `is`(expected))
+                }
+
+                @Test
+                @DisplayName("wildcard")
+                fun wildcard() {
+                    val expected = loadResource("tests/parser/marklogic-7.0/ElementDeclTest_Wildcard.txt")
+                    val actual = parseResource("tests/parser/marklogic-7.0/ElementDeclTest_Wildcard.xq")
+                    assertThat(prettyPrintASTNode(actual), `is`(expected))
+                }
+
+                @Test
+                @DisplayName("wildcard; compact whitespace")
+                fun wildcard_compactWhitespace() {
+                    val expected = loadResource("tests/parser/marklogic-7.0/ElementDeclTest_Wildcard_CompactWhitespace.txt")
+                    val actual = parseResource("tests/parser/marklogic-7.0/ElementDeclTest_Wildcard_CompactWhitespace.xq")
+                    assertThat(prettyPrintASTNode(actual), `is`(expected))
+                }
+            }
         }
 
         @Nested